Escolher versiculo part3

+++ CamadaDTO/clLivro.cs
@@ -2,7 +2,7 @@
 {
 	public class clLivro
 	{
-		public int IDLivro { get; set; }
+		public byte IDLivro { get; set; }
 		public string Livro { get; set; }
 		public string Abreviacao { get; set; }
 		public byte Capitulos { get; set; }
